+++ src/components/common/MonsterCardList/index.tsx
@@ -1,5 +1,5 @@
 import { Monster } from '@/model/monster';
-import styles from './css/MonsterCardList.module.css';
+import styles from './css/index.module.css';
 import MonsterCard from '../MonsterCard';
 import { Suspense } from 'react';
